user16

+++ app/src/main/java/cookbook/controller/CommentController.java
@@ -3,10 +3,11 @@ package cookbook.controller;
 import java.sql.SQLException;
 import java.text.SimpleDateFormat;
 
-import com.mysql.cj.xdevapi.Session;
-
 import cookbook.model.Comment;
 import cookbook.model.QueryMaker;
+import cookbook.model.Session;
+import cookbook.model.User;
+import cookbook.view.DisplayRecipeScene;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -32,36 +33,80 @@ public class CommentController {
   private Label myusername;
 
   private Comment comment; // Assume Comment class represents a comment object
+  private User user;
 
-  private AnchorPane parentAnchorPane;
+  private DisplayRecipeScene drsController;
 
-  public void setData(Comment comment, AnchorPane parent) {
-    String username = myusername.getText();
-    this.comment = comment;
-    parentAnchorPane = parent;
-    this.mycomment.setText(comment.getComment_text());
-    SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
-    myDate.setText(dateFormat.format(comment.getDate()));
-  }
+  public void setData(Comment comment, AnchorPane parent, DisplayRecipeScene drsController) {
+    this.drsController = drsController;
+    int commentId = comment.getId();
 
-  @FXML
-  void editComment(ActionEvent event) throws SQLException {
-    int userid = comment.getUser_id();
-
-    QueryMaker qm = new QueryMaker();
-    qm.editComment(comment);
-
-  }
-
-  @FXML
-  void removeComment(ActionEvent event) {
     try {
       QueryMaker qm = new QueryMaker();
-      qm.deleteComment(comment);
+      User commentUser = qm.retrieveCommentUser(commentId);
+
+      //String username = commentUser.getUsername();
+      String username = commentUser.getFname() + " " + commentUser.getLname();
+      myusername.setText(username);
+
     } catch (SQLException e) {
       System.out.println("Error: " + e.getMessage());
     }
 
+    this.comment = comment;
+
+    this.mycomment.setText(comment.getComment_text());
+    SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+
+    myDate.setText(dateFormat.format(comment.getDate()));
+    int userid = comment.getUser_id();
+
+    if (userid != Session.getCurrentUser().getUserId()) {
+      editButton.setVisible(false);
+      deleteButton.setVisible(false);
+      mycomment.setDisable(true);
+
+    }
+  }
+
+  @FXML
+  void editComment(ActionEvent event) throws SQLException {
+    String updatedComment = mycomment.getText();
+    int userid = comment.getUser_id();
+
+    if (userid == Session.getCurrentUser().getUserId()) {
+
+      try {
+
+        comment.setComment_text(updatedComment);
+        QueryMaker qm = new QueryMaker();
+        qm.editComment(comment);
+      } catch (SQLException e) {
+
+      }
+
+    } else {
+      System.out.println("only the creator of the comment can edit it.");
+    }
+
+    drsController.reloadComments();
+  }
+
+  @FXML
+  void removeComment(ActionEvent event) {
+    int userid = comment.getUser_id();
+    if (userid == Session.getCurrentUser().getUserId()) {
+      try {
+        QueryMaker qm = new QueryMaker();
+        qm.deleteComment(comment);
+      } catch (SQLException e) {
+        System.out.println("Error: " + e.getMessage());
+      }
+    } else {
+      System.out.println("only the creator of the comment can delete it");
+    }
+
+    drsController.reloadComments();
   }
 
 }
